+++ src/db/db_instance.rs
@@ -14,14 +14,13 @@ impl DbInstance {
     }
 
     pub async fn get_table_names(&self) -> Vec<String> {
-        let mut result = Vec::new();
-
         let read_access = self.tables.read().await;
-        for table_name in read_access.keys() {
-            result.push(table_name.to_string());
-        }
 
-        return result;
+        return read_access
+            .values()
+            .into_iter()
+            .map(|table| table.name.clone())
+            .collect();
     }
 
     pub async fn get_tables(&self) -> Vec<Arc<DbTable>> {
