+++ src/http/controllers/home_controller/mod.rs
@@ -1,0 +1,3 @@
+mod index_action;
+
+pub use index_action::IndexAction;
