+++ src/http/controllers/logs_controller/mod.rs
@@ -1,0 +1,7 @@
+mod by_process_action;
+mod by_table_action;
+mod home_action;
+mod logs;
+pub use by_process_action::LogsByProcessAction;
+pub use by_table_action::LogsByTableAction;
+pub use home_action::HomeAction;
