+++ src/http/controllers/mod.rs
@@ -1,8 +1,12 @@
 pub mod api;
+mod as_html;
 pub mod builder;
 pub mod bulk;
 pub mod gc;
+pub mod home_controller;
+pub mod logs_controller;
 pub mod multipart;
+pub mod prometheus_controller;
 pub mod row_controller;
 pub mod rows_controller;
 pub mod status;
