+++ src/http/controllers/prometheus_controller/mod.rs
@@ -1,0 +1,2 @@
+mod metrics_action;
+pub use metrics_action::MetricsAction;
